+++ cdp-agentkit-core/python/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - Added `wrap_eth` action to wrap ETH to WETH on Base.
+- Added support for Allora Chain machine learning (`get_price_prediction` and `get_all_topics` actions).
 
 ## [0.0.7] - 2025-01-08